Extract concatenated CharSequence to a separate file
